Move `listeners` to the `models` package
