+++ curated_transformers/tests/models/test_listeners.py
@@ -4,7 +4,7 @@ from spacy.tokens import Doc
 from spacy.training import Example
 from thinc.api import Config, reduce_mean
 
-from curated_transformers.listener import build_transformer_layers_listener_v1
+from curated_transformers.models.listener import build_transformer_layers_listener_v1
 from curated_transformers._compat import has_hf_transformers
 
 cfg_string_transformer_layers_listener = """
