+++ setup.cfg
@@ -29,9 +29,9 @@ spacy_architectures =
     curated-transformers.XLMRTransformer.v1 = curated_transformers.models:build_xlmr_transformer_model_v1
     curated-transformers.WithStridedSpans.v1 = curated_transformers.models:build_with_strided_spans_v1
     curated-transformers.ScalarWeight.v1 = curated_transformers.models:build_scalar_weight_v1
-    curated-transformers.TransformerLayersListener.v1 = curated_transformers.listener:build_transformer_layers_listener_v1
-    curated-transformers.LastTransformerLayerListener.v1 = curated_transformers.listener:build_last_transformer_layer_listener_v1
-    curated-transformers.ScalarWeightingListener.v1 = curated_transformers.listener:build_scalar_weighting_listener_v1
+    curated-transformers.TransformerLayersListener.v1 = curated_transformers.models.listener:build_transformer_layers_listener_v1
+    curated-transformers.LastTransformerLayerListener.v1 = curated_transformers.models.listener:build_last_transformer_layer_listener_v1
+    curated-transformers.ScalarWeightingListener.v1 = curated_transformers.models.listener:build_scalar_weighting_listener_v1
 
 spacy_callbacks =
     curated-transformers.gradual_transformer_unfreezing.v1 = curated_transformers.util:create_gradual_transformer_unfreezing
